Moved TG events to a separate unit
